Refactor clint init code
