+++ src/device.rs
@@ -1,1 +1,2 @@
 pub mod uart;
+pub mod clint;
