+++ src/hw/riscv.rs
@@ -28,29 +28,6 @@ pub const SIE_SEIE: u64 = 1 << 9; // external
 pub const SIE_STIE: u64 = 1 << 5; // timer
 pub const SIE_SSIE: u64 = 1 << 1; // software
 
-// CLINT := Core local interruptor (where the timer is).
-// CLINT_BASE: usize = 0x2000000; // clint is at this location in memlayout.
-// xv6-riscv C code:
-// #define CLINT_MTIMECMP(hartid) (CLINT + 0x4000 + 8*(hartid))
-// #define  CLINT_MTIME (CLINT + 0xBFF8) // cycles since boot.
-// int interval = 1000000; // cycles; about 1/10th second in qemu.
-// *(uint64*)CLINT_MTIMECMP(id) = *(uint64*)CLINT_MTIME + interval;
-
-// Need to write a value to the CLINT memory location.
-// This is mmio, as such there are safety concerns:
-//      https://doc.rust-lang.org/std/ptr/fn.write_volatile.html
-// 
-// Generate a machine lvl interrupt by setting mtime to be >= mtimecmp.
-pub fn write_clint(hartid: u64, base: usize, interval: u64) {
-    // Ok, treat base addr as a pointer we can write to.
-    let base = (base + 0x4000 + 8 * (hartid as usize)) as *mut u64;
-    unsafe {
-        base.write_volatile(base as u64 + 0xBFF8 + interval);
-    }
-}
-
-
-
 // Return id of current hart.
 // the "m" in "mstatus" means machine mode.
 // Note mstatus -> sstatus reg for supervisor mode.
